change function lua to use the new jemalloc alloc
