+++ src/eval.c
@@ -175,61 +175,6 @@ int luaRedisReplicateCommandsCommand(lua_State *lua) {
     return 1;
 }
 
-#if defined(USE_JEMALLOC)
-/* When lua uses jemalloc, pass in luaAlloc as a parameter of lua_newstate. */
-static void *luaAlloc(void *ud, void *ptr, size_t osize, size_t nsize) {
-    UNUSED(osize);
-
-    unsigned int tcache = (unsigned int)(uintptr_t)ud;
-    if (nsize == 0) {
-        zfree_with_flags(ptr, MALLOCX_ARENA(server.lua_arena) | MALLOCX_TCACHE(tcache));
-        return NULL;
-    } else {
-        return zrealloc_with_flags(ptr, nsize, MALLOCX_ARENA(server.lua_arena) | MALLOCX_TCACHE(tcache));
-    }
-}
-
-/* Create a lua interpreter, and use jemalloc as lua memory allocator. */
-lua_State *createLuaState(void) {
-    /* Every time a lua VM is created, a new private tcache is created for use.
-     * This private tcache will be destroyed after the lua VM is closed. */
-    unsigned int tcache;
-    size_t sz = sizeof(unsigned int);
-    int err = je_mallctl("tcache.create", (void *)&tcache, &sz, NULL, 0);
-    if (err) {
-        serverLog(LL_WARNING, "Failed creating the lua jemalloc tcache.");
-        exit(1);
-    }
-
-    /* We pass tcache as ud so that it is not bound to the server. */
-    return lua_newstate(luaAlloc, (void *)(uintptr_t)tcache);
-}
-
-/* Under jemalloc we need to create a new arena for lua to avoid blocking
- * defragger. */
-void scriptingSetup(void) {
-    unsigned int arena;
-    size_t sz = sizeof(unsigned int);
-    int err = je_mallctl("arenas.create", (void *)&arena, &sz, NULL, 0);
-    if (err) {
-        serverLog(LL_WARNING, "Failed creating the lua jemalloc arena.");
-        exit(1);
-    }
-    server.lua_arena = arena;
-}
-
-#else
-
-/* Create a lua interpreter and use glibc (default) as lua memory allocator. */
-lua_State *createLuaState(void) {
-    return lua_open();
-}
-
-/* There is nothing to set up under glib. */
-void scriptingSetup(void) {}
-
-#endif
-
 /* Initialize the scripting environment.
  *
  * This function is called the first time at server startup with
@@ -243,10 +188,8 @@ void scriptingSetup(void) {}
 void scriptingInit(int setup) {
     if (setup) {
         lctx.lua_client = NULL;
-        server.lua_arena = UINT_MAX;
         server.script_disable_deny_script = 0;
         ldbInit();
-        scriptingSetup();
     }
 
     lua_State *lua = createLuaState();
