+++ src/function_lua.c
@@ -423,7 +423,7 @@ static int luaRegisterFunction(lua_State *lua) {
 /* Initialize Lua engine, should be called once on start. */
 int luaEngineInitEngine(void) {
     luaEngineCtx *lua_engine_ctx = zmalloc(sizeof(*lua_engine_ctx));
-    lua_engine_ctx->lua = lua_open();
+    lua_engine_ctx->lua = createLuaState();
 
     luaRegisterRedisAPI(lua_engine_ctx->lua);
